requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
asgiref==3.8.1
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto3==1.36.6
botocore==1.36.6
certifi==2024.12.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==44.0.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2
django-filter==24.3
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-oss-storage==1.1.1
django-storages==1.14.4
djangorestframework==3.15.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.1.5
jmespath==0.10.0
jsonfield==3.1.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
mysqlclient==2.2.7
netifaces==0.11.0
oauthlib==3.2.0
oss2==2.19.1
packaging==21.3
pexpect==4.8.0
psycopg==3.2.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycryptodome==3.21.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymemcache==4.0.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
python-memcached==1.62
pytz==2024.2
PyYAML==5.4.1
redis==5.2.1
requests==2.32.3
s3transfer==0.11.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.17.0
sos==4.7.2
sqlparse==0.5.3
ssh-import-id==5.11
swapper==1.4.0
systemd-python==234
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
wrapt==1.17.2
zipp==1.0.0
zope.interface==5.4.0
